refactor: authCheck 파일 명 getAuthCheck으로 변경 #1144
