+++ frontend/src/common/apis/getAuthCheck.ts
